move client defaults to separate module
